+++ app/Data/ContactData.php
@@ -4,7 +4,12 @@ namespace App\Data;
 
 class ContactData
 {
+    const STATUS_INITIAL = 'Initial';
     const STATUS_READ = 'Read';
     const STATUS_UNREAD = 'Unread';
     const STATUS_ARCHIVED = 'Archived';
+
+    // Category Id
+    const CATEGORY_ID_PRIVATE = 1;
+    const CATEGORY_ID_GROUP = 2;
 }