+++ routes/web.php
@@ -13,6 +13,10 @@ Route::get('emails/baseReviewRequest', function () {
     return view("emails/baseReviewRequest");
 });
 
+Route::get('emails/join-conversation-email', function () {
+    return view("emails/joinConversationEmail");
+});
+
 Route::get('images/{size}/{filename}', function ($size, $filename) {
     return response()->file(storage_path('app/public/images/' . $size . '/' . $filename));
 })->name("images");