+++ CHANGELOG.md
@@ -4,6 +4,10 @@
 - Add column message_category_id at messages table
 - php artisan update:db
 - create chat_settings table
+- create contact_categories, groups table
+- add column contact_category_id at contact table
+- make nullable connection_user_id at contacts table
+- create group_invitations table